test: move test into the root folder
