+++ test/autoCleanup.js
@@ -1,5 +1,5 @@
 import { useEffect } from 'react'
-import { renderHook } from '../'
+import { renderHook } from '../src'
 
 // This verifies that by importing RHTL in an
 // environment which supports afterEach (like Jest)
